Add temp_dir
This temporary directory is used for creating a mini project within my project with its own package.json.
I used this to create a firebase server so that I could access data in my firebase database. As this is data collected from the field it is more usefully for testing than any I could cook up.